move to openapi structure
