+++ bruno/api/createUser.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: {{Localhost}}:8082/users
+  url: {{localhost}}:8082/api/{{version}}/users
   body: json
   auth: none
 }
