+++ bruno/api/folder.bru
@@ -1,0 +1,12 @@
+meta {
+  name: api
+  seq: 2
+}
+
+auth {
+  mode: inherit
+}
+
+vars:pre-request {
+  version: v1
+}
