+++ bruno/api/getUser--id.bru
@@ -1,11 +1,11 @@
 meta {
   name: getUser/:id
   type: http
-  seq: 2
+  seq: 1
 }
 
 get {
-  url: {{Localhost}}:8082/users/1
+  url: {{localhost}}:8082/api/{{version}}/users/1
   body: none
   auth: none
 }
