+++ bruno/api/getUsers.bru
@@ -1,11 +1,11 @@
 meta {
   name: getUsers
   type: http
-  seq: 3
+  seq: 1
 }
 
 get {
-  url: {{Localhost}}:8082/users
+  url: {{localhost}}:8082/api/{{version}}/users
   body: none
   auth: none
 }
