+++ bruno/collection.bru
@@ -6,7 +6,3 @@ meta {
 auth {
   mode: none
 }
-
-vars:pre-request {
-  Localhost: http://localhost
-}
