+++ bruno/environments/local.bru
@@ -1,0 +1,3 @@
+vars {
+  localhost: http://localhost
+}
