+++ internal/infrastructure/http/openapi/v1/handler/user/create_user.go
@@ -3,8 +3,8 @@ package user
 import (
 	"net/http"
 
-	"github.com/umefy/go-web-app-template/internal/infrastructure/http/handler/user/mapping"
 	"github.com/umefy/go-web-app-template/internal/infrastructure/http/middleware"
+	"github.com/umefy/go-web-app-template/internal/infrastructure/http/openapi/v1/handler/user/mapping"
 	api "github.com/umefy/go-web-app-template/openapi/protogen/v1/models"
 	"github.com/umefy/go-web-app-template/pkg/validation"
 	"github.com/umefy/godash/jsonkit"
