+++ internal/infrastructure/http/openapi/v1/handler/user/mapping/user_mapping.go
