+++ openapi/docs/api.yaml
@@ -12,10 +12,8 @@ info:
     name: MIT License
     identifier: MIT
 servers:
-  - url: https://api.server.test
-    description: Production server
-  - url: https://api.server.test
-    description: Production server
+  - url: http://localhost:8080/api/v1
+    description: Local server
 
 paths:
   /users:
